requirements.txt
gradio
sqlalchemy
langchain
langchain
langchain-community
psycopg2
langchain-openai
python-dotenv
langchain-ollama